chatbot/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
build==1.4.0
cachetools==6.2.4
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.4.1
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
courlan==1.3.2
cryptography==46.0.4
cssselect==1.4.0
dataclasses-json==0.6.7
dateparser==1.3.0
distro==1.9.0
durationpy==0.10
ecdsa==0.19.1
et_xmlfile==2.0.0
faiss-cpu==1.13.2
fastapi==0.128.0
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.20.3
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
htmldate==1.9.4
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jusText==3.0.2
jwt==1.4.0
kubernetes==35.0.0
langchain==1.2.6
langchain-chroma==1.1.0
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.7
langchain-openai==1.1.7
langchain-text-splitters==1.1.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.2
lxml==6.0.2
lxml_html_clean==0.4.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
networkx==3.6.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==2.4.1
oauthlib==3.3.1
onnxruntime==1.23.2
openai==2.15.0
openpyxl==3.1.5
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
ormsgpack==1.12.2
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
posthog==5.4.0
propcache==0.4.1
protobuf==6.33.3
pyarrow==22.0.0
pyasn1==0.6.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pypdfium2==5.3.0
PyPika==0.50.0
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.45
starlette==0.50.0
streamlit==1.52.2
streamlit-cookies-manager==0.2.0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.1
tokenizers==0.22.2
toml==0.10.2
torch==2.9.1
tornado==6.5.4
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.57.3
typer==0.21.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
watchdog==6.0.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
